fix: update prism-common-libs dependency to v1.0.1 and refactor config usage in file service
